Moving summary.dart to model.dart
