+++ lib/pana.dart
@@ -5,10 +5,10 @@
 export 'src/code_problem.dart';
 export 'src/fitness.dart';
 export 'src/license.dart';
+export 'src/model.dart';
 export 'src/package_analyzer.dart';
 export 'src/pkg_resolution.dart';
 export 'src/platform.dart';
 export 'src/pubspec.dart';
 export 'src/sdk_env.dart';
-export 'src/summary.dart';
 export 'src/utils.dart';
