+++ lib/src/mini_sum.dart
@@ -2,7 +2,7 @@ import 'dart:collection';
 import 'dart:convert';
 
 import 'code_problem.dart';
-import 'summary.dart';
+import 'model.dart';
 import 'utils.dart';
 
 class MiniSum {
